# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Management API",
        "description": "RESTful API for Risk Management System",
        "contact": {
            "email": "admin@riskmanagement.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/reports/dashboard": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get dashboard summary data",
                "operationId": "57f39aecded089a4dccaa4a7475929c8",
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reports/risk-summary": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get risk summary statistics",
                "operationId": "db025d23047c715b0ee662d3927b98c8",
                "responses": {
                    "200": {
                        "description": "Risk summary retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reports/risk-matrix": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get risk matrix data for visualization",
                "operationId": "e93efc0fdb7f7d769e17f79d2d36e617",
                "responses": {
                    "200": {
                        "description": "Risk matrix data retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reports/risks-by-category": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get risks breakdown by category",
                "operationId": "881b8d35c7f68376795ae5efb9140041",
                "responses": {
                    "200": {
                        "description": "Risks by category retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reports/risks-by-department": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get risks breakdown by department",
                "operationId": "bc215d8720c579d7b9bfc000344495e4",
                "responses": {
                    "200": {
                        "description": "Risks by department retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reports/overdue-actions": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get overdue mitigation actions",
                "operationId": "70c5c97b5b5716e532a3ccd3d3dc50e7",
                "responses": {
                    "200": {
                        "description": "Overdue actions retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reports/high-risk-items": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get high risk items requiring attention",
                "operationId": "dd1c68baed01b049ea9dbdb018ab7de4",
                "responses": {
                    "200": {
                        "description": "High risk items retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}